Send createFolders:true option to JSZip
The zip files generated by gulp-zip do not actually contain "real"
folders (that is, folders with the `D` attribute and `store` method).
Opening a zip with a GUI decompression utility shows folders, but
they're really "virtual" folders that are just defined by the path of
the files in the zip.  Extracting the zips creates the actual folder
structure, but a decompression utility looking specifically for folders
won't find anything, as the `D` attribute is not present.

By sending {createFolders: true} as an option to JSZip's `.file` method,
true folders are created (with D attributes, method properties, etc) rather
than just virtual representations of folders that are part of the files'
paths.

This requires updating the required version of JSZip to 2.4.0.